+++ repyability/tests/test_rbd_koon.py
@@ -23,6 +23,14 @@ def test_rbd_koon_k_given(rbd1_koon: RBD):
     assert rbd1_koon.G.nodes["sink"]["k"] == 1
 
 
+def test_rbd_graph_is_koon_rbd_false(rbd_series: RBD):
+    assert not rbd_series.G.is_koon_rbd()
+
+
+def test_rbd_graph_is_koon_rbd_true(rbd1_koon: RBD):
+    assert rbd1_koon.G.is_koon_rbd()
+
+
 def test_rbd_get_min_path_sets_rbd_series_koon(rbd_series_koon: RBD):
     assert set() == rbd_series_koon.get_min_path_sets()
 
@@ -175,3 +183,59 @@ def test_rbd_get_min_path_sets_rbd_double_parallel_koon(
         frozenset(["s", 1, 2, 3, 4, "t"]),
         frozenset(["s", 1, 2, 3, 5, "t"]),
     } == rbd.get_min_path_sets()
+
+
+def test_rbd_get_min_cut_sets_rbd_series_koon(rbd_series_koon: RBD):
+    # It's a series RBD with the middle node k=2 so the system doesn't work
+    # at all
+    assert {} == rbd_series_koon.get_min_cut_sets()
+
+
+def test_rbd_get_min_cut_sets_rbd1_koon(rbd1_koon: RBD):
+    assert {
+        frozenset(["pump1"]),
+        frozenset(["pump2"]),
+        frozenset(["valve"]),
+    } == rbd1_koon.get_min_cut_sets()
+
+
+def test_rbd_get_min_cut_sets_rbd2_koon(rbd2_koon: RBD):
+    assert {
+        frozenset([2]),
+        frozenset([3]),
+        frozenset([4]),
+        frozenset([5]),
+        frozenset([6]),
+        frozenset([7]),
+    } == rbd2_koon.get_min_cut_sets()
+
+
+def test_rbd_get_min_cut_sets_rbd3_koon1(rbd3_koon1: RBD):
+    assert {
+        frozenset([2, 4]),
+        frozenset([1, 4]),
+        frozenset([2, 3]),
+    } == rbd3_koon1.get_min_cut_sets()
+
+
+def test_rbd_get_min_cut_sets_rbd3_koon2(rbd3_koon2: RBD):
+    assert {
+        frozenset([1, 3]),
+        frozenset([2, 4]),
+        frozenset([1, 4]),
+        frozenset([5, 4]),
+        frozenset([1, 3]),
+        frozenset([5, 3]),
+        frozenset([2, 3]),
+    } == rbd3_koon2.get_min_cut_sets()
+
+
+def test_rbd_get_min_cut_sets_rbd3_koon3(rbd3_koon3: RBD):
+    assert {
+        frozenset([2, 4]),
+        frozenset([2, 1]),
+        frozenset([3]),
+        frozenset([1, 4]),
+        frozenset([4, 5]),
+        frozenset([4, 1]),
+    } == rbd3_koon3.get_min_cut_sets()
